requirements.txt
torch
torchaudio
librosa
soundfile
numpy
hydra-core
omegaconf
tqdm
wandb
matplotlib
scipy
gdown
